fixed bug when getting pizza: fixed by change type int => int64. Changing types in models and manager

+++ internal/model/pizza.go
@@ -1,12 +1,12 @@
 package model
 
 type Pizza struct {
-	ID         int    `json:"id" db:"id"`
-	ImageURL   string `json:"imageURL" db:"image_url"`
-	Name       string `json:"name" db:"name"`
-	Types      []int  `json:"types" db:"types"`
-	Sizes      []int  `json:"sizes" db:"sizes"`
-	Price      int    `json:"price" db:"price"`
-	CategoryID int    `json:"category_id" db:"category_id"`
-	Rating     int    `json:"rating" db:"rating"`
+	ID         int64   `json:"id" db:"id"`
+	ImageURL   string  `json:"imageURL" db:"image_url"`
+	Name       string  `json:"name" db:"name"`
+	Types      []int64 `json:"types" db:"types"`
+	Sizes      []int64 `json:"sizes" db:"sizes"`
+	Price      int64   `json:"price" db:"price"`
+	CategoryID int64   `json:"category_id" db:"category_id"`
+	Rating     int64   `json:"rating" db:"rating"`
 }
