+++ pkg/auth/token-manager.go
@@ -1,7 +1,7 @@
 package auth
 
 type TokenManager interface {
-	GenerateJWT(id int, role int) (string, error)
+	GenerateJWT(id int64, role int64) (string, error)
 	Parse(accessToken string) (interface{}, error)
 	RefreshJWT() (string, error)
 }
